requirements.txt
google-cloud-aiplatform[adk,agent_engines]==1.110.0
google-api-python-client
google-auth
google-auth-httplib2
